restore: set permission to the logs folder
This ensures to keep the right ownership for the logs directory.
After the app is removed, the log directory could have been kept. But the user does not exist anymore, and the owner is referenced by a user ID.
It could happen that the user ID has been assigned to another user in the meantime, and thus the log directory granted to that user.

So when the app is restored, the folder exists but it is not owned by the right user.